work model update

+++ app/django/work/models.py
@@ -375,7 +375,7 @@ class Issue(models.Model):
                                verbose_name='상위 업무', related_name='children')
     watchers = models.ManyToManyField('accounts.User', blank=True, verbose_name='업무 열람 공유자', related_name='watchers')
     is_private = models.BooleanField('비공개', default=False)
-    estimated_hours = models.PositiveSmallIntegerField('추정 소요시간', null=True, blank=True)
+    estimated_hours = models.DecimalField('추정 소요시간', max_digits=5, decimal_places=2, null=True, blank=True)
     start_date = models.DateField('시작 일자', null=True, blank=True)
     due_date = models.DateField('완료 기한', null=True, blank=True)
     PROGRESS_RATIO = (
